pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "stock-analysis"
version = "0.1.0"
description = "Fetch stock close and options chain data from free endpoints"
requires-python = ">=3.11"
dependencies = [
  "requests>=2.31.0",
  "python-dateutil>=2.9.0.post0",
  "fastapi>=0.110.0",
  "uvicorn[standard]>=0.27.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
